replace TrieDBMut with MemoryTrieDB, move node tests to node module
